# VEconomy-swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "VEconomy",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/security/login": {
      "get": {
        "tags": [
          "Security"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoggedUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoggedUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoggedUser"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Security"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoggedUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoggedUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoggedUser"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/security/changePass": {
      "put": {
        "tags": [
          "Security"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePass"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePass"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePass"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePass"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/security/user": {
      "post": {
        "tags": [
          "Security"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Security"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Security"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "string"
              }
            },
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/security/userPassword": {
      "put": {
        "tags": [
          "Security"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Credentials": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "nullable": true
          },
          "pass": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ChangePass": {
        "type": "object",
        "properties": {
          "oldPass": {
            "type": "string",
            "nullable": true
          },
          "pass": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LoggedUser": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "rights": {
            "$ref": "#/components/schemas/Rights"
          }
        },
        "additionalProperties": false
      },
...
      "Rights": {
        "enum": [
          1,
          2,
          4,
          8,
          16,
          32,
          64,
          128,
          256,
          378,
          481
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "rights": {
            "$ref": "#/components/schemas/Rights"
          },
          "validFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "validTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}